+++ source/edgevis.cpp
@@ -168,25 +168,25 @@ int body(ProgramOptionVariables pov)
     int c = 0;
     int spaceEdge;
     std::string name;
+    edgevis::EdgeVisibility Evis(edgemesh);
+    Evis.set_visual_mesh(geomMeshPoly);
+
     for (Edge e : edgemesh.mesh_edges){
-        if(c % 28 == 0) {
+        if(c % 1 == 0) {
             name = "images/" + pov.input_map_name + "_" + std::to_string(c) + ".pdf";
             std::cout << name << std::endl ;
             spaceEdge = c;
             r_v.clear(); l_v.clear(); v.clear();
-            if(c==-1){
-                std::cout << "calculating right \n";
-                r_v = edgevis::find_visibility(spaceEdge, edgemesh, true, geomMeshPoly, true);
-                std::cout << "calculating left \n";
-                l_v = edgevis::find_visibility(spaceEdge, edgemesh, false, geomMeshPoly, true);
-                v.reserve( r_v.size() + l_v.size() ); // preallocate memory
-            }else{
-                std::cout << "calculating right \n";
-                r_v = edgevis::find_visibility(spaceEdge, edgemesh, true, geomMeshPoly, false);
-                std::cout << "calculating left \n";
-                l_v = edgevis::find_visibility(spaceEdge, edgemesh, false, geomMeshPoly, false);
-                v.reserve( r_v.size() + l_v.size() ); // preallocate memory
-            }
+            if(c==-1)
+                Evis.switch_debug(true);
+            else
+                Evis.switch_debug(false);
+            std::cout << "calculating right \n";
+            r_v = Evis.find_visibility(spaceEdge, true);
+            std::cout << "calculating left \n";
+            l_v = Evis.find_visibility(spaceEdge, false);
+            v.reserve( r_v.size() + l_v.size() ); // preallocate memory
+
             v.insert( v.end(), r_v.begin(), r_v.end() );
             v.insert( v.end(), l_v.begin(), l_v.end() );
             std::cout << "visualizing\n";