+++ source/libs/edgevis/include/edgevis/search/expansion.h
@@ -11,7 +11,7 @@ bool is_on_segment(Point A, Point B, Point C);
 
 /*
  * side means on which side of edge we want to look since we want to calculate visibility separately.
- * lets say true means right and false means left
+ * let say true mean child_R and false means child_L
  */
 int get_edge_init_nodes(Edge edge, bool side, const Mesh& mesh, SearchNode* initNodes);
 
@@ -27,13 +27,17 @@ bool is_observable(Point left_parent,
 /*
  *
  */
-std::vector<int> normalise_ids(std::vector<int> elements, int new_start, int& offset);
+int normalise(const Polygon& P, int transition_R, std::vector<int>* sorted_vertices, std::vector<int>* sorted_polygons);
+
+int find_visible(const Mesh& mesh, SearchNode& node, std::vector<int>& sorted_vertices, int* right_visible, int* left_visible);
 
 /*
  *
  */
-void recompute_roots(SearchNode &node, int num);
+void recompute_roots(SearchNode &node);
 
+
+SearchNode init_temp_node(SearchNode& node);
 /*
  *
  */